% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkd.R
\name{dLKD}
\alias{dLKD}
\title{Distribuicao Katz Lagrangiana (LKD)}
\usage{
dLKD(x, a = 5, b = 0.02, beta = 0.3, log = FALSE, ...)
}
\arguments{
\item{x}{quantis de probabilidade da LKD}

\item{a}{parametro 1}

\item{b}{parametro 2}

\item{beta}{parametro 3, tambem de nome beta}

\item{log}{se TRUE retorna o log das probabilidades}

\item{...}{passagem de argumentos}
}
\description{
Probabilidades acumuladas da LKD.
}
\details{
Para lidar com fatoriais envolvendo números decimais utilizamos a função Gamma, além disso, visando resolver problemas de números muito grandes nos fatoriais, usamos a aproximação de Stirling.
}
\examples{
dLKD(x = 0:10, a = 5, b = 0.05, beta = 0.1)
}

