% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkd.R
\name{est_mle_v2}
\alias{est_mle_v2}
\title{Estimativas de Maxima log-verossimilhança da lkd em outra abordagem}
\usage{
est_mle_v2(y)
}
\description{
Estimativas de Maxima log-verossimilhança da lkd em outra abordagem
}
