% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkd.R
\name{is.bn}
\alias{is.bn}
\title{Dados os parametros verifica se a lkd é uma Binomial(N, p = beta)}
\usage{
is.bn(n, a, b, beta, c = NULL)
}
\description{
Dados os parametros verifica se a lkd é uma Binomial(N, p = beta)
}
