% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkd.R
\name{is.lkd}
\alias{is.lkd}
\title{Dados os parametros verifica se é lkd}
\usage{
is.lkd(n, a, b, beta, c = NULL)
}
\arguments{
\item{n}{número de observacções desejadas da lkd(a, b, beta)}

\item{a}{parametro 1}

\item{b}{parametro 2}

\item{beta}{parametro 3}

\item{c}{parametro 3}
}
\description{
Dados os parametros verifica se é lkd
}
