% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkd.R
\name{is.nbn}
\alias{is.nbn}
\title{Dados os parametros verifica se a lkd é uma Binomial Negativa(N, p = beta)}
\usage{
is.nbn(n, a, b, beta, c = NULL)
}
\description{
Dados os parametros verifica se a lkd é uma Binomial Negativa(N, p = beta)
}
