% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkd.R
\name{is.po}
\alias{is.po}
\title{Dados os parametros verifica se a lkd é uma Poisson(theta = a)}
\usage{
is.po(n, a, b, beta, c = NULL)
}
\description{
Dados os parametros verifica se a lkd é uma Poisson(theta = a)
}
