% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkd.R
\name{lkd_est_mle}
\alias{lkd_est_mle}
\title{Estimativas de Maxima log-verossimilhança da lkd}
\usage{
lkd_est_mle(y, a, beta, ...)
}
\arguments{
\item{y}{vetor de amostras de uma lkd}

\item{a}{parametro 1}

\item{...}{passagem de argumentos}

\item{b}{parametro 2}
}
\description{
Estimativas de Maxima log-verossimilhança da lkd
}
