% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkd.R
\name{lkd_est_mzcf}
\alias{lkd_est_mzcf}
\title{Estimativas com base em momento e zero-class para a lkd}
\usage{
lkd_est_mzcf(y, a, beta)
}
\arguments{
\item{y}{vetor de amostras de uma lkd}

\item{a}{parametro 1}

\item{b}{parametro 2}
}
\description{
Estimativas com base em momento e zero-class para a lkd
}
