% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkd.R
\name{lkd_gradient_log_vero_consul}
\alias{lkd_gradient_log_vero_consul}
\title{Função Gradiente analítica}
\usage{
lkd_gradient_log_vero_consul(pars, ...)
}
\arguments{
\item{pars}{vetor de parametros}

\item{...}{passagem de argumentos}
}
\value{
vetor das tres derivadas de primeira ordem
}
\description{
calcula analiticamente o gradiente da função de log-verossimilhança presente em Consul e Famoye (2016). Por definição, a função gradiente é formada por todas a as derivadas analílicas de primeira ordem da função de log-verossimilhança.  Esta função é pré-compilada para tornar sua execução mais rápida. Ela consome um vetor de observações de uma lkd(a, b, beta) de nome y que deve estar no ambiente do R.
}
