% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkd.R
\name{lkd_hessian_log_vero_consul}
\alias{lkd_hessian_log_vero_consul}
\title{Função Hessiana analítica}
\usage{
lkd_hessian_log_vero_consul(pars, ...)
}
\arguments{
\item{pars}{vetor de parametros}

\item{...}{passagem de argumentos}
}
\value{
matriz 3 x 3.
}
\description{
calcula analiticamente o Hessiano da função de log-verossimilhança presente em Consul e Famoye (2016). Por definição, a matriz Hessiana é a matrix das derivadas analílicas de segunda ordem da função de log-verossimilhança.  Esta função é pré-compilada para tornar sua execução mais rápida. Ela consome um vetor de observações de uma lkd(a, b, beta) de nome y que deve estar no ambiente do R.
}
