% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkd.R
\name{lkd_log_vero_consul}
\alias{lkd_log_vero_consul}
\title{Função de log-verosimilhança}
\usage{
lkd_log_vero_consul(pars, ...)
}
\arguments{
\item{pars}{vetor de parametros}

\item{...}{passagem de argumentos}
}
\description{
calcula a função de log-verossimilhança presente em Consul e Famoye (2016).
Esta função é pré-compilada para tornar sua execução mais rápida. Ela consome um vetor
de observações de uma lkd(a, b, beta) de nome y que deve estar no ambiente do R.
}
