% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkd.R
\name{newton_raphson}
\alias{newton_raphson}
\title{Algoritmo Newton-Rapson matricial}
\usage{
newton_raphson(initial, escore, hessiano, tol = 1e-04, max.iter = 500,
  n.dim = 3, trace = TRUE, ...)
}
\arguments{
\item{initial}{valores do chute inicial}

\item{escore}{função escore da log-verossimilhança}

\item{hessiano}{função Hessiana da log-verossimilhança}

\item{tol}{tolerãncia para o erro entre uma estimativa e a anterior}

\item{max.iter}{numero de iterações alogortmo antes da convergência}

\item{n.dim}{número de dimensões (parãmetros) da log-verossimilhança}

\item{trace}{se TRUE esibe as estimativas em cada iteração}

\item{...}{outros parametros}
}
\value{
lista contendo dois elementos: o Hessiano da ultima iteração e as estimativas
}
\description{
Função generica para obter estimativas por Newton-Rapson usando o gradiente e o hessiano da função. Se o sistema dormado pelo Hessiano for singular, ou seja, sem solução. Usa-se a inversa de Moore-Pen Rose como aproximação para as estimativas.
}
\examples{
# N <- 500
# y <- 0:(N-1)
# a <- 10; b <- 0.5; beta <- 0.05
# pars <- c(a, b, beta)
# y <- rlkd(N, a, b, beta)
# hist(y)
# start <- unname(lkd_est_mzcf(y, a, beta))
# k <- newton_raphson(initial = start
# , escore = lkd_gradient_log_vero_consul
# , hessiano = lkd_hessian_log_vero_consul
# , max.iter = 5000, n.dim = 3, tol = 0.00001)
}
