% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkd.R
\name{plkd}
\alias{plkd}
\title{Distribuicao Katz Lagrangiana (LKD)}
\usage{
plkd(q, a = 5, b = 0.02, beta = 0.3, size = 1, f_bino = FALSE,
  log = FALSE, ...)
}
\arguments{
\item{q}{vetor de quantis de probabilidade da LKD}

\item{a}{parametro 1}

\item{b}{parametro 2}

\item{beta}{parametro 3, tambem de nome beta}

\item{size}{utlizado quando a lkd deriva Binomial ou Binomial Negativa}

\item{f_bino}{se TRUE retorna dlkd na forma fatorial, se FALSE retorna a forma função gamma.}

\item{log}{se TRUE retorna o log das probabilidades}

\item{...}{passagem de argumentos}
}
\value{
vetor de probabilidades acumuladas para cada quartil da lkd(a, b, beta)
}
\description{
Retorna probabilidades acumuladas de uma lkd(a, b, beta)
}
\details{
todo
}
\examples{
plkd(0:5, 2, -0.2, 0.2)
plkd(0:5, 2, -0.2, 0.2, log = TRUE)
plkd(0:5, 2, -0.2, 0.2, log = TRUE, f_bino = FALSE)
plkd(0:5, 2, 0, 0.2)
plkd(0:5, 2, 0.7, .9)
plot(function(x) plkd(x, 20, 0.5, 0.9), 0, 50, main = "LKD", type = "h")

}
