% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkd.R
\name{qlkd}
\alias{qlkd}
\title{Distribuicao Katz Lagrangiana (LKD)}
\usage{
qlkd(p, a = 5, b = 0.02, beta = 0.3, log = FALSE)
}
\arguments{
\item{p}{vetor de probabilidade da LKD}

\item{a}{parametro 1}

\item{b}{parametro 2}

\item{beta}{parametro 3, tambem de nome beta}
}
\description{
Inversa das probabilidades acumuladas da lkd(a, b, beta) obtida por simulação via método da transformação integral da probabilidade.
}
\details{
Gerar quantis da LKD utilizando a transformação integral de probabilidade através de aproximação numérica.
}
\examples{
p <- plkd(0:10, 2, 0.1, 0.2, log = FALSE)
qlkd(p, 2, 0.1, 0.2, log = FALSE)
qlkd(ppois(0:5, lambda = 2), 2, 0, 0, log = FALSE)
par(mfrow = c(1,2))
plot(function(x) plkd(x, 2, 0.1, 0.2), ylim = c(0,1), xlim = c(0, 15),
 main = "Acumulada LKD", type = "l")
plot(function(x) qlkd(x, 2, 0.1, 0.2), ylim = c(0,15), xlim = c(0, 1),
 main = "Inversa Acumulada LKD", type = "l")
par(mfrow = c(1,1))

}
