% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkd.R
\name{rlkd}
\alias{rlkd}
\title{Distribuicao Katz Lagrangiana (LKD)}
\usage{
rlkd(n, a = 5, b = 0.02, beta = 0.3, size = 1, log = FALSE, acum = FALSE)
}
\arguments{
\item{n}{número de observacções desejadas da lkd(a, b, beta)}

\item{a}{parametro 1}

\item{b}{parametro 2}

\item{beta}{parametro 3}

\item{size}{utlizado quando a lkd deriva Binomial ou Binomial Negativa}

\item{log}{se TRUE retorna o log das probabilidades}

\item{acum}{se TRUE retorna a probabilidade acumulada no quatil X, senão retorna X como uma realização da lkd(a, b, beta)}
}
\value{
vetor de observações de uma lkd(a, b, beta)
}
\description{
Geração de números aleatórios da LKD pelo método da inversa da função de probabilidade acumulada.
Ocasionalmente a lkd pode derivar outras ditribuições. Por hora temos: lkd, kd, binomial, binomial negativa e Poisson. Outras serão configuradas em outras versões. No caso de recair em uma distribuição conhecida, o algoritimo de geração de numeros aleatórios destas distribuições é utilizado. Vide \code{\link{rbinom}}, \code{\link{rnbinom}} e \code{\link{rpois}}. Demais casos recam no algoritimo desenvolvido com base em Consul e Famoye (2006).
}
\details{
Esta função utiliza uma função interna que determina amostras da lkd(a, b, beta) pelo método da inversa da função de probabilidade acumulada.
}
\examples{
rlkd(10, 2, -0.2, 0.2)
rlkd(10, 2, -0.2, 0.3)
rlkd(100, 2, 0, 0.2)
x <- rlkd(500, 2, 0.2, 0.3)
hist(x, main = 'Amostras da lkd(2, 0.2, 0.3)')
}
