% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkd.R
\name{dlkd}
\alias{dlkd}
\title{Distribuicao Katz Lagrangiana (LKD)}
\usage{
dlkd(x, a = 5, b = 0.02, beta = 0.3, size = 1, f_bino = FALSE,
  log = FALSE, ...)
}
\arguments{
\item{x}{um vetor de observações de uma lkd(a, b, beta)}

\item{a}{parametro 1}

\item{b}{parametro 2}

\item{beta}{parametro 3, tambem de nome beta}

\item{size}{utlizado quando a lkd deriva Binomial ou Binomial Negativa}

\item{f_bino}{se TRUE retorna dlkd na forma fatorial, se FALSE retorna a forma função gamma.}

\item{log}{se TRUE retorna o log das probabilidades}

\item{...}{passagem de argumentos}
}
\value{
vetor de probabilidades da lkd(a, b, beta)
}
\description{
Esta função retorna probabilidades da LKD. Podendo ser: lkd, kd, binomial, binomial negativa, binomial negativa generalizada, Poisson e Poisson Generalizada de acordo com os parãmetros a, b e beta.
}
\details{
Para lidar com fatoriais envolvendo números decimais utilizamos a função Gamma. Além disso, todas as contas são feitas em log e ao final retornam em exponencial.
}
\examples{
X <- 0:20
sum(dlkd(X, 2, 0.1, 0.2, log = FALSE))
sum(dlkd(X, 2, 0, 0.2))
sum(dlkd(X, 2, 0.7, 0.9, log = FALSE, f_bino = FALSE))
sum(dlkd(X, 2, 0, 0, log = FALSE))
plot(function(x) dlkd(x, 20, 0.5, 0.9), 0, 50, main = "LKD", type = "h")

}
