% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkd.R
\name{is.gnbn}
\alias{is.gnbn}
\title{Dados os parametros verifica se a lkd é uma Binomial Negativa Generalizada(N, beta, a)}
\usage{
is.gnbn(n, a, b, beta, c = NULL)
}
\description{
Dados os parametros verifica se a lkd é uma Binomial Negativa Generalizada(N, beta, a)
}
