% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkd.R
\name{lkd_est_mzcf}
\alias{lkd_est_mzcf}
\title{Estimativas com bese em momemento e zero-class para a lkd}
\usage{
lkd_est_mzcf(y, a, beta, ...)
}
\description{
Estimativas com bese em momemento e zero-class para a lkd
}
