% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkd.R
\name{pLKD}
\alias{pLKD}
\title{Distribuicao Katz Lagrangiana (LKD)}
\usage{
pLKD(q, a = 5, b = 0.02, beta = 0.3, log = F, ...)
}
\arguments{
\item{q}{quantis de probabilidade da LKD}

\item{a}{parametro 1}

\item{b}{parametro 2}

\item{beta}{parametro 3, tambem de nome beta}

\item{log}{se TRUE retorna o log das probabilidades}

\item{...}{passagem de argumentos}
}
\description{
Probabilidades acumuladas da LKD.
}
\details{
todo
}
\examples{
pLKD(q = 50, a = 5, b = 0.05, beta = 0.1)
}
