% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkd.R
\name{qLKD}
\alias{qLKD}
\title{Distribuicao Katz Lagrangiana (LKD)}
\usage{
qLKD(p, a = 5, b = 0.02, beta = 0.3, ...)
}
\arguments{
\item{p}{vetor de probabilidade da LKD}

\item{a}{parametro 1}

\item{b}{parametro 2}

\item{beta}{parametro 3, tambem de nome beta}

\item{...}{passagem de argumentos}
}
\description{
Probabilidades acumuladas da LKD.
}
\details{
Gerar quantis da LKD utilizando a transformação integral de probabilidade através de aproximação numérica.
}
\examples{
qLKD(10, a = 5, b = 0.05, beta = 0.1)
}

