% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lkd.R
\name{rLKD}
\alias{rLKD}
\title{Distribuicao Katz Lagrangiana (LKD)}
\usage{
rLKD(n, a = 5, b = 0.02, beta = 0.3, ...)
}
\arguments{
\item{n}{número de observacções}

\item{a}{parametro 1}

\item{b}{parametro 2}

\item{beta}{parametro 3, tambem de nome beta}

\item{...}{passagem de argumentos}
}
\description{
Geração de números aleatórios da LKD pelo método da inversa da função de probabilidade acumulada.
}
\examples{
rLKD(50, a = 5, b = 0.05, beta = 0.1)
}
